ssifi/requirements.txt
appdirs==1.4.4
asttokens==2.0.5
audioread==2.1.9
backcall==0.2.0
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///C:/ci/click_1646056799533/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///C:/ci/cryptography_1639472366776/work
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
fonttools==4.33.2
huggingface-hub==0.5.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1648544472910/work
ipython==8.2.0
jedi==0.18.1
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
kiwisolver==1.4.2
librosa==0.9.1
llvmlite==0.38.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
numba==0.55.1
numpy @ file:///C:/ci/numpy_and_numpy_base_1649782971440/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
pooch==1.6.0
prompt-toolkit==3.0.29
protobuf==3.19.1
pure-eval==0.2.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.12.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
PyYAML==6.0
regex @ file:///C:/ci/regex_1648447923622/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
resampy==0.2.2
sacremoses==0.0.43
scikit-learn==1.0.2
scipy==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
stack-data==0.2.0
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.11.0
torchvision==0.12.0
tqdm @ file:///C:/ci/tqdm_1650618053064/work
traitlets==5.1.1
transformers==4.11.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///C:/ci/urllib3_1650640043075/work
wcwidth==0.2.5
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
